refactor: 패키지 분리 및 user 도메인 하위 도메인들의 위치 변경
