+++ app/src/main/java/com/joonhyeok/app/user/application/AccountService.java
@@ -4,9 +4,9 @@ import com.joonhyeok.app.user.application.dto.AccountBalanceQuery;
 import com.joonhyeok.app.user.application.dto.AccountBalanceQueryResult;
 import com.joonhyeok.app.user.application.dto.AccountChargeCommand;
 import com.joonhyeok.app.user.application.dto.AccountChargeResult;
-import com.joonhyeok.app.user.domain.Account;
-import com.joonhyeok.app.user.domain.User;
-import com.joonhyeok.app.user.domain.UserRepository;
+import com.joonhyeok.app.user.domain.user.Account;
+import com.joonhyeok.app.user.domain.user.User;
+import com.joonhyeok.app.user.domain.user.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
