+++ app/src/main/java/com/joonhyeok/app/user/application/UserPayService.java
@@ -8,7 +8,11 @@ import com.joonhyeok.app.reservation.domain.Reservation;
 import com.joonhyeok.app.reservation.domain.ReservationRepository;
 import com.joonhyeok.app.user.application.dto.UserPayCommand;
 import com.joonhyeok.app.user.application.dto.UserPayResult;
-import com.joonhyeok.app.user.domain.*;
+import com.joonhyeok.app.user.domain.pay.PayEvent;
+import com.joonhyeok.app.user.domain.pay.PayEventPublisher;
+import com.joonhyeok.app.user.domain.user.PayValidator;
+import com.joonhyeok.app.user.domain.user.User;
+import com.joonhyeok.app.user.domain.user.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
