+++ app/src/main/java/com/joonhyeok/app/user/application/outbox/OutboxSendFailCommand.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.infra.domain.payEvent;
+package com.joonhyeok.app.user.application.outbox;
 
 public record OutboxSendFailCommand(
         String type,
