+++ app/src/main/java/com/joonhyeok/app/user/domain/outbox/OutboxStatus.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.infra.domain.payEvent;
+package com.joonhyeok.app.user.domain.outbox;
 
 public enum OutboxStatus {
     INIT, SEND_FAIL, SEND_SUCCESS
