+++ app/src/main/java/com/joonhyeok/app/user/domain/outbox/PayOutboxListener.java
@@ -1,0 +1,7 @@
+package com.joonhyeok.app.user.domain.outbox;
+
+import com.joonhyeok.app.user.domain.pay.PayEvent;
+
+public interface PayOutboxListener {
+    public void listen(PayEvent payEvent);
+}
