+++ app/src/main/java/com/joonhyeok/app/user/domain/pay/PayEventListener.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.domain;
+package com.joonhyeok.app.user.domain.pay;
 
 public interface PayEventListener {
     public void sendPayInfo(PayEvent payEvent);
