+++ app/src/main/java/com/joonhyeok/app/user/domain/pay/PayEventPublisher.java
@@ -1,4 +1,4 @@
-package com.joonhyeok.app.user.domain;
+package com.joonhyeok.app.user.domain.pay;
 
 public interface PayEventPublisher {
     public void publish(PayEvent payEvent);
