+++ app/src/test/java/com/joonhyeok/app/queue/QueueControllerIntegrateTest.java
@@ -3,9 +3,9 @@ package com.joonhyeok.app.queue;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.joonhyeok.app.queue.domain.QueueRepository;
 import com.joonhyeok.app.queue.domain.QueueStatus;
-import com.joonhyeok.app.user.domain.Account;
-import com.joonhyeok.app.user.domain.User;
-import com.joonhyeok.app.user.domain.UserRepository;
+import com.joonhyeok.app.user.domain.user.Account;
+import com.joonhyeok.app.user.domain.user.User;
+import com.joonhyeok.app.user.domain.user.UserRepository;
 import com.joonhyeok.openapi.models.RegisterQueueRequest;
 import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import org.junit.jupiter.api.Test;
