+++ app/src/test/java/com/joonhyeok/app/user/PayValidatorTest.java
@@ -4,9 +4,9 @@ import com.joonhyeok.app.concert.domain.Seat;
 import com.joonhyeok.app.concert.domain.SeatStatus;
 import com.joonhyeok.app.reservation.domain.Reservation;
 import com.joonhyeok.app.reservation.domain.ReservationStatus;
-import com.joonhyeok.app.user.domain.Account;
-import com.joonhyeok.app.user.domain.PayValidator;
-import com.joonhyeok.app.user.domain.User;
+import com.joonhyeok.app.user.domain.user.Account;
+import com.joonhyeok.app.user.domain.user.PayValidator;
+import com.joonhyeok.app.user.domain.user.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
