+++ app/src/test/java/com/joonhyeok/app/user/UserPayServiceIntegrateTest.java
@@ -9,8 +9,16 @@ import com.joonhyeok.app.reservation.domain.ReservationStatus;
 import com.joonhyeok.app.user.application.UserPayService;
 import com.joonhyeok.app.user.application.dto.UserPayCommand;
 import com.joonhyeok.app.user.application.dto.UserPayResult;
-import com.joonhyeok.app.user.domain.*;
-import com.joonhyeok.app.user.infra.domain.payEvent.*;
+import com.joonhyeok.app.user.infra.domain.outbox.PayOutboxKafkaListener;
+import com.joonhyeok.app.user.domain.pay.PayEvent;
+import com.joonhyeok.app.user.domain.pay.PayEventListener;
+import com.joonhyeok.app.user.domain.user.Account;
+import com.joonhyeok.app.user.domain.user.User;
+import com.joonhyeok.app.user.domain.user.UserRepository;
+import com.joonhyeok.app.user.domain.outbox.Outbox;
+import com.joonhyeok.app.user.application.outbox.OutboxFindCommand;
+import com.joonhyeok.app.user.application.OutboxService;
+import com.joonhyeok.app.user.domain.outbox.OutboxStatus;
 import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import jakarta.persistence.EntityNotFoundException;
 import org.junit.jupiter.api.DisplayName;
